requirements.txt
streamlit
opencv-python-headless
pandas
numpy
scikit-learn
joblib
Pillow
